Refactored and deleted useless classes
