--- src/main/java/com/example/lotteon/repository/product/ShoppingCartRepository.java
@@ -1,9 +1,0 @@
-package com.example.lotteon.repository.product;
-
-import com.example.lotteon.entity.product.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ShoppingCartRepository  extends JpaRepository<Product, Integer> {
-}
